changed2

+++ src/main/java/school/mjc/stage0/module3/task4/FormulaCreator.java
@@ -3,7 +3,7 @@ package school.mjc.stage0.module3.task4;
 public class FormulaCreator {
     public static void main(String[] args) {
         int a = 1, b = 10;
-        float formula = (9 * a * a - 5 * b + 2 + a - 7)*((a + b - 4 * a * b) / 2);
+        double formula = (9 * a * a - 5 * b + 2 + a - 7)*((a + b - 4 * a * b) / 2);
 
     }
 }
